Refactored and actually got proper tokens, no commas or semis tho
