+++ program.java
@@ -1,7 +1,7 @@
 CLASS Pgm1 {
     CONST M = 13, N = 56;
     VAR X, Y, Z;
-
+    /* hello */
     Y = 97;
     X = M * N + 18 - y;
 }
