+++ src/compiler.rs
@@ -1,2 +1,1 @@
 pub mod lexanalysis;
-pub mod fsa;
