+++ src/main.rs
@@ -5,9 +5,9 @@ pub use crate::compiler::lexanalysis;
 fn main() {
     // NOTE: init state table in main instead?
     let mut lex = lexanalysis::Tokenize::read_file("program.java").unwrap();
-    
-    println!("{:?}",lex.next());
-    println!("{:?}",lex.next());
-    println!("{:?}",lex.next());
-    println!("{:?}",lex.next());
+
+    // Loop through all tokens until the Scanner returns None
+    while let Some(token) = lex.next() {
+        println!("{}", token.unwrap());
+    }
 }
